+++ libindi/drivers/telescope/telescope_simulator.h
@@ -31,7 +31,7 @@ public:
 
     virtual const char *getDefaultName();
     virtual bool Connect();
-    virtual bool Connect(const char *port, uint16_t baud);
+    virtual bool Connect(const char *port, uint32_t baud);
     virtual bool Disconnect();
     virtual bool ReadScopeStatus();
     virtual bool initProperties();
